Extracts AgedBrie into a separate class
