+++ Java/com/gildedrose/GildedRose.java
@@ -55,60 +55,6 @@ class GildedRose {
         return item.sellIn;
     }
 
-    private class AgedBrieItem {
-
-        private boolean qualityIsLessThanMaximum(Item currentItem) {
-            return getQualityOf(currentItem) < MAXIMUM_QUALITY;
-        }
-
-        private int incrementQualityOfItem(Item currentItem) {
-            return getQualityOf(currentItem) + 1;
-        }
-
-        private void qualityIncreasesAsSellInDecreases(Item currentItem) {
-            defaultQualityIncrease(currentItem);
-            defaultSellInDecrement(currentItem);
-        }
-
-        private void defaultQualityIncrease(Item currentItem) {
-            setItemQualityTo(currentItem, incrementQualityOfItem(currentItem));
-        }
-
-        private void setItemQualityTo(Item currentItem, int quality) {
-            currentItem.quality = quality;
-        }
-
-        private void brieIncreasesQualityWithAge(Item currentItem) {
-            if (getSellInOf(currentItem) < 0) {
-                defaultQualityIncrease(currentItem);
-            }
-        }
-
-        private void process(Item currentItem) {
-            if (qualityIsLessThanMaximum(currentItem)) {
-                qualityIncreasesAsSellInDecreases(currentItem);
-                brieIncreasesQualityWithAge(currentItem);
-            }
-        }
-
-        private int getQualityOf(Item item) {
-            return item.quality;
-        }
-
-        private void defaultSellInDecrement(Item currentItem) {
-            setSellInOf(currentItem, decrementSellInOfItem(currentItem));
-        }
-
-        private void setSellInOf(Item currentItem, int sellIn) {
-            currentItem.sellIn = sellIn;
-        }
-
-        private int decrementSellInOfItem(Item currentItem) {
-            return getSellInOf(currentItem) - 1;
-        }
-
-    }
-
     private class BackstagePassItem {
 
         private void qualityIncreasesAsSellInDecreases(Item currentItem) {
